removed older version of tokens.py, changed tests
