+++ app/tokens.py
@@ -1,9 +1,7 @@
-import time
 from concurrent import futures
 from typing import Optional, Sequence
 
-from hexbytes import HexBytes
-from web3.types import BlockData
+from web3.types import BlockData, TxData
 
 from app.crud import Crud
 from app.web3_client import Web3Client
@@ -13,32 +11,25 @@ def query_ERC20_tokens(*, w3: Web3Client, crud: Crud) -> None:
     block = _get_first_block(w3, crud)
     crud.set_is_block_fetch(True)
     last_block_hash = crud.get_last_block_hash()
-    count = 0
-    first_start = time.time()
     while block and block["hash"] != last_block_hash:
-        start = time.time()
         crud.set_current_block(block["hash"])
-        transactions: Sequence[HexBytes] = block["transactions"]  # type: ignore
+        transactions: Sequence[TxData] = block["transactions"]  # type: ignore
         _find_contract_creations(w3, crud, transactions)
-        block = w3.get_parent_block(block)
-        end = time.time()
-        count += 1
-        print("block num:", count)
-        print("block time:", end - start)
-        print("total time:", end - first_start)
+        block = w3.get_parent_block(block, full_transactions=True)
+
     _finish_process(crud)
 
 
 def _get_first_block(w3: Web3Client, crud: Crud) -> Optional[BlockData]:
     start_block_hash = crud.get_start_block_hash()
     if not start_block_hash:
-        block = w3.get_latest_block()
+        block = w3.get_latest_block(full_transactions=True)
         crud.set_start_block(block["hash"])
         return block
     else:
         current_block_hash = crud.get_current_block_hash()
         assert current_block_hash is not None
-        return w3.get_block_by_hash(current_block_hash)
+        return w3.get_block_by_hash(current_block_hash, full_transactions=True)
 
 
 def _finish_process(crud: Crud) -> None:
@@ -48,8 +39,8 @@ def _finish_process(crud: Crud) -> None:
     crud.set_is_block_fetch(False)
 
 
-def _find_contract_creations(
-    w3: Web3Client, crud: Crud, transactions: Sequence[HexBytes]
+def _find_contract_creations_with_threads(
+    w3: Web3Client, crud: Crud, transactions: Sequence[TxData]
 ) -> None:
     if len(transactions) < 1:
         return
@@ -59,13 +50,16 @@ def _find_contract_creations(
         _ = executor.map(lambda p: _save_if_erc20_token(*p), args)
 
 
-def _save_if_erc20_token(
-    w3: Web3Client, crud: Crud, transaction_hash: HexBytes
+def _find_contract_creations(
+    w3: Web3Client, crud: Crud, transactions: Sequence[TxData]
 ) -> None:
-    if not w3.is_transaction_contract_creation(transaction_hash):
+    for t in transactions:
+        _save_if_erc20_token(w3, crud, t)
+
+
+def _save_if_erc20_token(w3: Web3Client, crud: Crud, transaction: TxData) -> None:
+    if not w3.is_transaction_contract_creation(transaction):
         return
-    print("found a contract")
-    address = w3.get_contract_address_by_transaction_hash(transaction_hash)
+    address = w3.get_contract_address_by_transaction_hash(transaction["hash"])
     if w3.is_contract_erc20(address):
-        print("found a token")
         crud.save_token_address(address)
